refactor(editor): Move htmlUtils into @n8n/composables behind re-export shim (no-changelog)
Relocates htmlUtils (incl. sanitizeHtml/sanitizeIfString) and its two
sanitization constants (ALLOWED_HTML_TAGS/ALLOWED_HTML_ATTRIBUTES, used
only by htmlUtils) into @n8n/composables via git mv. The old @/app paths
become thin re-export shims carrying deprecation comments; call sites are
untouched. Mirrors the useClipboard/injectionKeys move (#34229).

Part of CAT-3686 Stage 3 (N8N-36). useMessage is deferred to a follow-up:
it renders N8nCheckbox from @n8n/design-system, which already depends on
@n8n/composables, so moving it here too would create a build cycle.
